# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ParkingZoneDemandLevel =
  (typeof ParkingZoneDemandLevel)[keyof typeof ParkingZoneDemandLevel];

export const ParkingZoneDemandLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type ParkingZoneZoneType =
  (typeof ParkingZoneZoneType)[keyof typeof ParkingZoneZoneType];

export const ParkingZoneZoneType = {
  street: "street",
  mall: "mall",
  office: "office",
  residential: "residential",
  mixed: "mixed",
} as const;

export interface ParkingZone {
  id: number;
  name: string;
  description?: string;
  lat: number;
  lng: number;
  totalSpots: number;
  availableSpots: number;
  demandLevel: ParkingZoneDemandLevel;
  zoneType: ParkingZoneZoneType;
  /** 0-100 AI confidence score for availability */
  confidenceScore: number;
  pricePerHour?: number;
  createdAt?: string;
}

export type CreateParkingZoneBodyZoneType =
  (typeof CreateParkingZoneBodyZoneType)[keyof typeof CreateParkingZoneBodyZoneType];

export const CreateParkingZoneBodyZoneType = {
  street: "street",
  mall: "mall",
  office: "office",
  residential: "residential",
  mixed: "mixed",
} as const;

export interface CreateParkingZoneBody {
  name: string;
  description?: string;
  lat: number;
  lng: number;
  totalSpots: number;
  zoneType: CreateParkingZoneBodyZoneType;
  pricePerHour?: number;
}

export type ParkingSpotStatus =
  (typeof ParkingSpotStatus)[keyof typeof ParkingSpotStatus];

export const ParkingSpotStatus = {
  available: "available",
  occupied: "occupied",
  unknown: "unknown",
} as const;

export interface ParkingSpot {
  id: number;
  zoneId: number;
  spotNumber: string;
  status: ParkingSpotStatus;
  lastReportedAt?: string;
  confidenceScore: number;
  reportCount: number;
}

export type CreateParkingSpotBodyStatus =
  (typeof CreateParkingSpotBodyStatus)[keyof typeof CreateParkingSpotBodyStatus];

export const CreateParkingSpotBodyStatus = {
  available: "available",
  occupied: "occupied",
  unknown: "unknown",
} as const;

export interface CreateParkingSpotBody {
  zoneId: number;
  spotNumber: string;
  status?: CreateParkingSpotBodyStatus;
}

export type UpdateParkingSpotBodyStatus =
  (typeof UpdateParkingSpotBodyStatus)[keyof typeof UpdateParkingSpotBodyStatus];

export const UpdateParkingSpotBodyStatus = {
  available: "available",
  occupied: "occupied",
  unknown: "unknown",
} as const;

export interface UpdateParkingSpotBody {
  status?: UpdateParkingSpotBodyStatus;
}

export type ParkingReportReportType =
  (typeof ParkingReportReportType)[keyof typeof ParkingReportReportType];

export const ParkingReportReportType = {
  leaving: "leaving",
  arriving: "arriving",
  available: "available",
  occupied: "occupied",
} as const;

export interface ParkingReport {
  id: number;
  spotId?: number;
  zoneId: number;
  userId?: number;
  reportType: ParkingReportReportType;
  isVerified: boolean;
  pointsEarned: number;
  createdAt: string;
}

export type CreateParkingReportBodyReportType =
  (typeof CreateParkingReportBodyReportType)[keyof typeof CreateParkingReportBodyReportType];

export const CreateParkingReportBodyReportType = {
  leaving: "leaving",
  arriving: "arriving",
  available: "available",
  occupied: "occupied",
} as const;

export interface CreateParkingReportBody {
  spotId?: number;
  zoneId: number;
  userId?: number;
  reportType: CreateParkingReportBodyReportType;
}

export type ParkingPredictionConfidence =
  (typeof ParkingPredictionConfidence)[keyof typeof ParkingPredictionConfidence];

export const ParkingPredictionConfidence = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface ParkingPrediction {
  zoneId: number;
  zoneName?: string;
  timeSlot: string;
  /** 0-100 probability of finding parking */
  availabilityProbability: number;
  predictedAvailableSpots?: number;
  confidence: ParkingPredictionConfidence;
  dayOfWeek?: string;
}

export type BestArrivalTimeAlternativeTimesItem = {
  time?: string;
  probability?: number;
};

export interface BestArrivalTime {
  zoneId: number;
  bestTime: string;
  probabilityAtBestTime: number;
  alternativeTimes: BestArrivalTimeAlternativeTimesItem[];
  tip?: string;
}

export type UserRank = (typeof UserRank)[keyof typeof UserRank];

export const UserRank = {
  newcomer: "newcomer",
  contributor: "contributor",
  trusted: "trusted",
  expert: "expert",
  legend: "legend",
} as const;

export interface User {
  id: number;
  username: string;
  totalPoints: number;
  totalReports: number;
  accuracyRate: number;
  /** 0-100 trust score based on report accuracy */
  trustScore: number;
  rank: UserRank;
  /** Total monetary rewards in INR */
  rewardsEarned: number;
  createdAt?: string;
}

export interface CreateUserBody {
  username: string;
  email: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  totalPoints: number;
  totalReports: number;
  accuracyRate: number;
  badge: string;
}

export type HeatmapPointDemandLevel =
  (typeof HeatmapPointDemandLevel)[keyof typeof HeatmapPointDemandLevel];

export const HeatmapPointDemandLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface HeatmapPoint {
  zoneId: number;
  zoneName?: string;
  lat: number;
  lng: number;
  demandLevel: HeatmapPointDemandLevel;
  /** 0-100 percentage */
  occupancyRate: number;
  timeOfDay?: string;
  dayOfWeek?: string;
}

export interface AnalyticsSummary {
  totalZones: number;
  totalSpots: number;
  totalAvailableSpots: number;
  totalReportsToday: number;
  totalActiveUsers: number;
  averageConfidenceScore: number;
  highDemandZones: number;
  lowDemandZones: number;
}

export type PeakHourDataDemandLevel =
  (typeof PeakHourDataDemandLevel)[keyof typeof PeakHourDataDemandLevel];

export const PeakHourDataDemandLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface PeakHourData {
  hour: number;
  label: string;
  occupancyRate: number;
  demandLevel: PeakHourDataDemandLevel;
  zoneId?: number;
}

export type ListZonesParams = {
  lat?: number;
  lng?: number;
  radius?: number;
};

export type ListSpotsParams = {
  zoneId?: number;
  available?: boolean;
};

export type ListReportsParams = {
  spotId?: number;
  zoneId?: number;
  limit?: number;
};

export type ListPredictionsParams = {
  zoneId?: number;
  targetTime?: string;
  dayOfWeek?: string;
};

export type GetBestArrivalTimeParams = {
  zoneId: number;
  date?: string;
};

export type GetLeaderboardParams = {
  limit?: number;
};

export type GetHeatmapParams = {
  timeOfDay?: string;
  dayOfWeek?: string;
};

export type GetPeakHoursParams = {
  zoneId?: number;
};