# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Parking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /spots:
    get:
      operationId: listSpots
      tags: [spots]
      summary: List parking spots
      parameters:
        - name: zoneId
          in: query
          schema:
            type: integer
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of parking spots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingSpot"
    post:
      operationId: createSpot
      tags: [spots]
      summary: Create a parking spot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateParkingSpotBody"
      responses:
        "201":
          description: Created spot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingSpot"
  /spots/{id}:
    get:
      operationId: getSpot
      tags: [spots]
      summary: Get spot by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Parking spot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingSpot"
    patch:
      operationId: updateSpot
      tags: [spots]
      summary: Update a parking spot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateParkingSpotBody"
      responses:
        "200":
          description: Updated spot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingSpot"
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List crowdsourced reports
      parameters:
        - name: spotId
          in: query
          schema:
            type: integer
        - name: zoneId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingReport"
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a crowdsourced parking report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateParkingReportBody"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingReport"
  /predictions:
    get:
      operationId: listPredictions
      tags: [predictions]
      summary: Get AI predictions for parking availability
      parameters:
        - name: zoneId
          in: query
          schema:
            type: integer
        - name: targetTime
          in: query
          schema:
            type: string
        - name: dayOfWeek
          in: query
          schema:
            type: string
      responses:
        "200":
          description: AI predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingPrediction"
  /predictions/best-time:
    get:
      operationId: getBestArrivalTime
      tags: [predictions]
      summary: Get the best arrival time for a zone
      parameters:
        - name: zoneId
          in: query
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Best arrival time recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BestArrivalTime"
...
  /analytics/heatmap:
    get:
      operationId: getHeatmap
      tags: [analytics]
      summary: Get demand heatmap data
      parameters:
        - name: timeOfDay
          in: query
          schema:
            type: string
        - name: dayOfWeek
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HeatmapPoint"
...
  /analytics/peak-hours:
    get:
      operationId: getPeakHours
      tags: [analytics]
      summary: Get peak hour demand data per zone
      parameters:
        - name: zoneId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Peak hours data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeakHourData"
components:
  schemas:
...
    ParkingSpot:
      type: object
      properties:
        id:
          type: integer
        zoneId:
          type: integer
        spotNumber:
          type: string
        status:
          type: string
          enum: [available, occupied, unknown]
        lastReportedAt:
          type: string
        confidenceScore:
          type: number
        reportCount:
          type: integer
      required:
        - id
        - zoneId
        - spotNumber
        - status
        - confidenceScore
        - reportCount
    CreateParkingSpotBody:
      type: object
      properties:
        zoneId:
          type: integer
        spotNumber:
          type: string
        status:
          type: string
          enum: [available, occupied, unknown]
      required:
        - zoneId
        - spotNumber
    UpdateParkingSpotBody:
      type: object
      properties:
        status:
          type: string
          enum: [available, occupied, unknown]
    ParkingReport:
      type: object
      properties:
        id:
          type: integer
        spotId:
          type: integer
        zoneId:
          type: integer
        userId:
          type: integer
        reportType:
          type: string
          enum: [leaving, arriving, available, occupied]
        isVerified:
          type: boolean
        pointsEarned:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - zoneId
        - reportType
        - isVerified
        - pointsEarned
        - createdAt
    CreateParkingReportBody:
      type: object
      properties:
        spotId:
          type: integer
        zoneId:
          type: integer
        userId:
          type: integer
        reportType:
          type: string
          enum: [leaving, arriving, available, occupied]
      required:
        - zoneId
        - reportType
    ParkingPrediction:
      type: object
      properties:
        zoneId:
          type: integer
        zoneName:
          type: string
        timeSlot:
          type: string
        availabilityProbability:
          type: number
          description: 0-100 probability of finding parking
        predictedAvailableSpots:
          type: integer
        confidence:
          type: string
          enum: [low, medium, high]
        dayOfWeek:
          type: string
      required:
        - zoneId
        - timeSlot
        - availabilityProbability
        - confidence
    BestArrivalTime:
      type: object
      properties:
        zoneId:
          type: integer
        bestTime:
          type: string
        probabilityAtBestTime:
          type: number
        alternativeTimes:
          type: array
          items:
            type: object
            properties:
              time:
                type: string
              probability:
                type: number
        tip:
          type: string
      required:
        - zoneId
        - bestTime
        - probabilityAtBestTime
        - alternativeTimes
...
    HeatmapPoint:
      type: object
      properties:
        zoneId:
          type: integer
        zoneName:
          type: string
        lat:
          type: number
        lng:
          type: number
        demandLevel:
          type: string
          enum: [low, medium, high]
        occupancyRate:
          type: number
          description: 0-100 percentage
        timeOfDay:
          type: string
        dayOfWeek:
          type: string
      required:
        - zoneId
        - lat
        - lng
        - demandLevel
        - occupancyRate
...
    PeakHourData:
      type: object
      properties:
        hour:
          type: integer
        label:
          type: string
        occupancyRate:
          type: number
        demandLevel:
          type: string
          enum: [low, medium, high]
        zoneId:
          type: integer
      required:
        - hour
        - label
        - occupancyRate
        - demandLevel