# lib/api-zod/src/generated/types/createParkingReportBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateParkingReportBodyReportType } from "./createParkingReportBodyReportType";

export interface CreateParkingReportBody {
  spotId?: number;
  zoneId: number;
  userId?: number;
  reportType: CreateParkingReportBodyReportType;
}