# lib/api-zod/src/generated/types/createParkingSpotBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateParkingSpotBodyStatus } from "./createParkingSpotBodyStatus";

export interface CreateParkingSpotBody {
  zoneId: number;
  spotNumber: string;
  status?: CreateParkingSpotBodyStatus;
}