# lib/api-zod/src/generated/types/parkingPrediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking API
 * OpenAPI spec version: 0.1.0
 */
import type { ParkingPredictionConfidence } from "./parkingPredictionConfidence";

export interface ParkingPrediction {
  zoneId: number;
  zoneName?: string;
  timeSlot: string;
  /** 0-100 probability of finding parking */
  availabilityProbability: number;
  predictedAvailableSpots?: number;
  confidence: ParkingPredictionConfidence;
  dayOfWeek?: string;
}