# lib/api-zod/src/generated/types/parkingSpot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking API
 * OpenAPI spec version: 0.1.0
 */
import type { ParkingSpotStatus } from "./parkingSpotStatus";

export interface ParkingSpot {
  id: number;
  zoneId: number;
  spotNumber: string;
  status: ParkingSpotStatus;
  lastReportedAt?: string;
  confidenceScore: number;
  reportCount: number;
}